requirements.txt
streamlit
python-dotenv
google-generativeai
flake8
pytest